+++ lib/mdxishTags.ts
@@ -7,19 +7,16 @@ import mdxishMdxComponentBlocks from '../processor/transform/mdxish/components/m
 import mdxishTables from '../processor/transform/mdxish/tables/mdxish-tables';
 import { isMDXElement } from '../processor/utils';
 
-import { jsxTableFromMarkdown } from './mdast-util/jsx-table';
-import { magicBlockFromMarkdown } from './mdast-util/magic-block';
-import { mdxComponentFromMarkdown } from './mdast-util/mdx-component';
-import { jsxTable } from './micromark/jsx-table';
-import { magicBlock } from './micromark/magic-block';
-import { mdxComponent } from './micromark/mdx-component';
+import { FEATURES, mdxishExtensions } from './micromark/mdxish-extensions';
+
+const { micromarkExtensions, fromMarkdownExtensions } = mdxishExtensions(FEATURES.tags);
 
 const tags = (doc: string) => {
   const set = new Set<string>();
 
   const processor = remark()
-    .data('micromarkExtensions', [jsxTable(), magicBlock(), mdxComponent()])
-    .data('fromMarkdownExtensions', [jsxTableFromMarkdown(), magicBlockFromMarkdown(), mdxComponentFromMarkdown()])
+    .data('micromarkExtensions', micromarkExtensions)
+    .data('fromMarkdownExtensions', fromMarkdownExtensions)
     .use(mdxishMdxComponentBlocks)
     .use(mdxishTables);
   const tree = processor.parse(doc);